+++ src/pages/interview/components/TravelType/styles.ts
@@ -2,11 +2,6 @@ import styled from "styled-components";
 
 export const FormStyled = styled.form`
 
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    height: 100%;
-
     .titleDestinyType {
         font-size: 24px;
         margin: -1vh 0 2vh 8vw;
